requirements.txt
Django==1.6
South==0.8.4
Whoosh==2.5.5
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-haystack==2.1.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
